Additional tests for getIn and hasIn
Factors some out of updateIn
